Address PR feedback 2nd round
