+++ docs/developer/notification-center.md
@@ -12,7 +12,7 @@ The Notification Center adds a 'bell' icon in the top-right of the Rancher Manag
 
 ## Adding a Notification
 
-> Note: Types are defined in `shell/notifications.ts`
+> Note: Types are defined in `types/notifications`
 
 There is a new VueX store with actions under the `notifications` store.
 
@@ -49,6 +49,8 @@ Example notification of type `Announcement`:
 ![Example Notification](./images/notification-example.png)
 
 Note that the `id` field is optional - if not specified, the notification will be assigned a unique identifier. Using a known identifier is only really required if you wish to be able to look up a notification by a known identifier OR you wish to be able yo update a notification after being added.
+
+> Note: Be careful not to include sensitive information in notifications, since these are persisted to the browser's local storage
 
 ### Actions
